Couple changes on the Cypress docs
Explicitly highlighted the npm i as I did not see it when it was in the sentence. Also switched to a different syntax on th egit clone commend - I see an error for it (which I could fix)... I think this one may work more universally (but I could be wrong).
I also missed the copying of the env file, so pulled it into a code block to make it obvious,.