Changes to Typescript QUickstart
A large number of our quick starts reference the docs in the readme rather than duplicating the steps to run - I changed the Readme to follow this convention and pointed users to it.

Placed the steps to run the quick start in the docs.

Removed material explaining how the tracetest agent connection, collector configuration, and docker compose network are configured so the quick start focuses solely on the use of the NPM package.